requirements.txt
Django==2.1.5
twilio==6.7.1
Pillow==3.4.1
django-celery-beat==1.1.1
django-celery-results==1.0.1
djangorestframework==3.8.2
phonenumbers==8.9.1
django-controlcenter==0.2.6
spacy==2.0.18
ephem==3.7.6
